Cast mode_t to unsigned int for formatted output.

+++ src/filemanager/chmod.c
@@ -183,7 +183,7 @@ chmod_callback (Dlg_head * h, Widget * sender, dlg_msg_t msg, int parm, void *da
         if (id >= 0)
         {
             c_stat ^= check_perm[id].mode;
-            g_snprintf (buffer, sizeof (buffer), "%o", c_stat);
+            g_snprintf (buffer, sizeof (buffer), "%o", (unsigned int) c_stat);
             label_set_text (statl, buffer);
             chmod_toggle_select (h, id);
             mode_change = 1;
@@ -351,7 +351,7 @@ chmod_cmd (void)
         add_widget (ch_dlg, label_new (FY + 6, FX + 2, c_fown));
         c_fgrp = str_trunc (get_group (sf_stat.st_gid), 21);
         add_widget (ch_dlg, label_new (FY + 8, FX + 2, c_fgrp));
-        g_snprintf (buffer, sizeof (buffer), "%o", c_stat);
+        g_snprintf (buffer, sizeof (buffer), "%o", (unsigned int) c_stat);
         statl = label_new (FY + 4, FX + 2, buffer);
         add_widget (ch_dlg, statl);
